Fix integer overflow in video frame stride/size computation
The per-plane stride was computed in int:

    stride[0] = (width * f.bytesPerSample + (alignment - 1)) & ~(alignment - 1);

with the only guard being the width <= 0 / height <= 0 check and no upper
bound. For a large script-supplied width (e.g. BlankClip(width=2^30,
format=RGBS), where width * 4 == 2^32) the product wraps int to ~0, so
the stride collapses while the frame still advertises the full width. The
plane buffer (new VSPlaneData(stride * height)) is then allocated far too
small and the filter writes width * height * bytesPerSample bytes into it
-- a heap overflow. The chroma stride additionally truncated through an
int plane23 temporary.

Compute the stride in ptrdiff_t and the allocation size in size_t in both
video-frame constructors. The maximum possible value, (2^31-1)^2 * 4, is
strictly below 2^64, so the widened size computation cannot overflow; an
oversized request now reaches the allocator at its true size instead of
silently wrapping to a tiny one.

+++ src/core/vscore.cpp
@@ -186,10 +186,10 @@ VSFrame::VSFrame(const VSVideoFormat &f, int width, int height, const VSFrame *p
     format.vf = f;
     numPlanes = format.vf.numPlanes;
 
-    stride[0] = (width * (f.bytesPerSample) + (alignment - 1)) & ~(alignment - 1);
+    stride[0] = (static_cast<ptrdiff_t>(width) * f.bytesPerSample + (alignment - 1)) & ~static_cast<ptrdiff_t>(alignment - 1);
 
     if (numPlanes == 3) {
-        int plane23 = ((width >> format.vf.subSamplingW) * (format.vf.bytesPerSample) + (alignment - 1)) & ~(alignment - 1);
+        ptrdiff_t plane23 = ((static_cast<ptrdiff_t>(width) >> format.vf.subSamplingW) * format.vf.bytesPerSample + (alignment - 1)) & ~static_cast<ptrdiff_t>(alignment - 1);
         stride[1] = plane23;
         stride[2] = plane23;
     } else {
@@ -197,9 +197,9 @@ VSFrame::VSFrame(const VSVideoFormat &f, int width, int height, const VSFrame *p
         stride[2] = 0;
     }
 
-    data[0] = new VSPlaneData(stride[0] * height, *core->memory);
+    data[0] = new VSPlaneData(static_cast<size_t>(stride[0]) * height, *core->memory);
     if (numPlanes == 3) {
-        size_t size23 = stride[1] * (height >> format.vf.subSamplingH);
+        size_t size23 = static_cast<size_t>(stride[1]) * (height >> format.vf.subSamplingH);
         data[1] = new VSPlaneData(size23, *core->memory);
         data[2] = new VSPlaneData(size23, *core->memory);
     }
@@ -220,10 +220,10 @@ VSFrame::VSFrame(const VSVideoFormat &f, int width, int height, const VSFrame * 
     format.vf = f;
     numPlanes = format.vf.numPlanes;
 
-    stride[0] = (width * (format.vf.bytesPerSample) + (alignment - 1)) & ~(alignment - 1);
+    stride[0] = (static_cast<ptrdiff_t>(width) * format.vf.bytesPerSample + (alignment - 1)) & ~static_cast<ptrdiff_t>(alignment - 1);
 
     if (numPlanes == 3) {
-        int plane23 = ((width >> format.vf.subSamplingW) * (format.vf.bytesPerSample) + (alignment - 1)) & ~(alignment - 1);
+        ptrdiff_t plane23 = ((static_cast<ptrdiff_t>(width) >> format.vf.subSamplingW) * format.vf.bytesPerSample + (alignment - 1)) & ~static_cast<ptrdiff_t>(alignment - 1);
         stride[1] = plane23;
         stride[2] = plane23;
     } else {
@@ -241,9 +241,9 @@ VSFrame::VSFrame(const VSVideoFormat &f, int width, int height, const VSFrame * 
             data[i]->add_ref();
         } else {
             if (i == 0) {
-                data[i] = new VSPlaneData(stride[i] * height, *core->memory);
+                data[i] = new VSPlaneData(static_cast<size_t>(stride[i]) * height, *core->memory);
             } else {
-                data[i] = new VSPlaneData(stride[i] * (height >> format.vf.subSamplingH), *core->memory);
+                data[i] = new VSPlaneData(static_cast<size_t>(stride[i]) * (height >> format.vf.subSamplingH), *core->memory);
             }
         }
     }
